requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
datasets>=2.14.0

# Computer Vision and Image Processing
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Hugging Face Datasets and API
datasets>=2.14.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data Analysis and Visualization
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
jupyter>=1.0.0

# Hugging Face Hub
huggingface-hub>=0.17.0